Greatly (~3x) increased range bonus for the final optics upgrade.
The goal of the range increase was always to reduce movement needed
so that the player can stand in one place and "draw" structures, but
the original maximum range wasn't enough to cover the area provided
by ultrawide displays or the use of 1x scaling.

It's still not quite far enough to fully cover 21:9 @ 1x scaling,
but it should be good enough for now.